Add option to turn off SD/MMC polling
This is the check for whether the SD card is removed. This is a
little silly when the SD card has the boot media. It doesn't really
cause a problem except that the disk activity light blinks after
every check and this bothers some people who think that the system
should be completely idle.